+++ test/nextgen/operators/matmul/pack_rhs/matmul_pack_rhs_wrapper_registry.hpp
@@ -109,6 +109,9 @@ create_matmul_rhs_pack_nxk_qsi4c32p16vsx4s4s0_qsu4c32_f32_bf16_sme();
 [[nodiscard]] bool is_shape_suitable_rhs_nxk_qsi4c32p16vsx4s4s0_qsu4c32_f32_bf16_sme(
     size_t shape_m, size_t shape_n, size_t shape_k, const MatrixPortion& portion);
 
+/// Creates a wrapper for kai_matmul_rhs_pack_nxk_qsi8cxp_qsi8cx_neon kernel.
+[[nodiscard]] std::unique_ptr<KernelWrapper<MatShape>> create_matmul_rhs_pack_nxk_qsi8cxp_qsi8cx_neon();
+
 /// Checks if the portion produces non-empty RHS packing tiles for the x32p4vsx1 KxN matmul operator.
 [[nodiscard]] bool is_shape_suitable_rhs_kxn_x32p4vsx1bx32_x32_x32_sme(
     size_t shape_m, size_t shape_n, size_t shape_k, const MatrixPortion& portion);
@@ -165,6 +168,10 @@ create_matmul_rhs_pack_nxk_qsi4c32p16vsx4s4s0_qsu4c32_f32_bf16_sme();
 
 /// Checks if the portion produces non-empty RHS packing tiles for the x16p16vsx2bx16 KxN matmul operator.
 [[nodiscard]] bool is_shape_suitable_rhs_kxn_x16p16vsx2bx16_x16_x16_sve(
+    size_t shape_m, size_t shape_n, size_t shape_k, const MatrixPortion& portion);
+
+/// Checks if the portion produces non-empty RHS packing tiles for the qsi8cxp_qsi8cx matmul operator.
+[[nodiscard]] bool is_shape_suitable_rhs_nxk_qsi8cxp_qsi8cx_neon(
     size_t shape_m, size_t shape_n, size_t shape_k, const MatrixPortion& portion);
 
 }  // namespace kai::test