+++ test/nextgen/reference/unary_elementwise.cpp
@@ -1,5 +1,6 @@
 //
 // SPDX-FileCopyrightText: Copyright 2025-2026 Arm Limited and/or its affiliates <open-source-office@arm.com>
+// SPDX-FileCopyrightText: Copyright 2026 Fujitsu Limited
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -93,6 +94,8 @@ UnaryElementwiseFn make_change_signedness(DataType dtype) {
         case DataType::U2:
         case DataType::I2:
             return unary_elementwise<ChangeSignednessOp<UInt2>>;
+        case DataType::I8:
+            return unary_elementwise<ChangeSignednessOp<int8_t>>;
 
         default:
             KAI_TEST_ERROR("Not supported.");